Fix SetKeyboardIndicatorsDataPDU: ledFlags is UINT16 not UINT32
MS-RDPBCGR §2.2.8.2.1.3.3.1 defines ledFlags as a 16-bit unsigned
integer (2 bytes).  The original implementation used uint32 (4 bytes),
causing binary.Read to attempt reading 6 bytes from a 4-byte PDU body,
resulting in 'unexpected EOF' errors that disrupted the RDP session.

+++ protocol/pdu/data.go
@@ -790,7 +790,7 @@ func (d *FontMapDataPDU) Unpack(r io.Reader) error {
 // MS-RDPBCGR 2.2.8.2.1.3.3.1
 type SetKeyboardIndicatorsDataPDU struct {
 	UnitId   uint16 `struc:"little"`
-	LedFlags uint32 `struc:"little"`
+	LedFlags uint16 `struc:"little"`
 }
 
 func (*SetKeyboardIndicatorsDataPDU) Type2() uint8 {
